requirements.txt
qrcode[pil]
pyzbar
pillow
pyotp
